Move 30FPS specific details
