+++ darksoulstas.py
@@ -3,14 +3,15 @@ Launch a Dark Souls TAS Console with helpful commands pre-loaded.
 
 Uses the STDLIB 'code' module to launch the terminal.
 """
-import sys
-import pydoc
 import code
+import pydoc
+import sys
 import textwrap
 
 from ds_tas import KeySequence, KeyPress, basics
-from ds_tas.scripts import menus, glitches, timers
 from ds_tas.engine import TAS
+from ds_tas.scripts import timers
+from ds_tas.scripts.ptde import glitches, menus
 
 __version__ = '3.0.0b2'
